# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API Documentation Taeyoung",
        "description": "API Documentation for Taeyoung",
        "version": "1.0.0"
    },
    "paths": {
...
        "/dgush-backend/public/api/province/{id}": {
            "get": {
                "tags": [
                    "Province"
                ],
                "summary": "Get province by id",
                "operationId": "eefc0a02b536f7de9011bf52ce7b53f5",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID of province",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ProvinceResource"
                                }
                            }
                        }
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "ProvinceResource": {
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": "1"
                    },
                    "name": {
                        "type": "string",
                        "example": "Jawa Barat"
                    }
                },
                "type": "object"
            },
...
        },
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "name": "Authorization",
                "in": "header",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}